requirements.txt
telethon
python-dotenv
pandas
transformers
datasets
torch
numpy